Start logging to new trails after DataResources
There is a short window after the trail is created in add_data_trails and before the EventSelectors are configured in update_trails where management events could be delivered to the trail. This moves the start_logging call after the EventSelectors have been set and IncludeManagementEvents is set to False.