Add EC2 action propagate-spot-tags.
This action can propagate either all, or a subset of, the tags
that are set at the Spot Instance Request.
It can helps for setting a tagging policy as for the moment spot
instance requests do not propagate tags to EC2 spot instances.
(While it seems to be the case on spot fleets, though)